gui: Preliminary support for DualSense haptics on Linux.
From my testing, haptics with PulseAudio are not working properly, so
using Pipewire as a backend is recommended (and picked by default, if
available via an SDL hint).

Support for macOS and Windows is still missing, but will most
probably only consist of expanding the DualSense audio device detection
logic to the device identifiers exposed by their respecive audio
backends.

The feature is currently hidden behind a setting or command-line flag
(`--dualsense`).