Initial version of HDR-capable Placebo Vulkan renderer.
This adds a second renderer besides the previous OpenGL renderer.
It uses libplacebo with its Vulkan output to draw the frames to the
stream window.

The renderer can be configured in the settings and still defaults to
OpenGL, since on my laptop Placebo is still dropping frames at 4k
rendering resolution on my laptop (Ryzen 7 6850U with Rembrandt
graphics), even on the fast preset. This is in contrast to the OpenGL
renderer, which keeps up nicely without dropping any frames.

Placebo offers three presets with increasing amounts of post-processing.
The highest preset e.g. implements debanding, which almost gets rid of
the annoying color banding in the PS5's main screen interface. When the
Placebo renderer is enabled, users can take their pick between the three
presets.

HDR is currently only implemented insofar as users can set the codec to
"H265 HDR" in the settings (only available when the Placebo renderer is
selected) which will output a 10bit image. This is however possibly
still missing any potential HDR metadata from the stream, which remains
an area of active research (due  to missing hardware on my end) and will
hopefully be confirmed to be working already or be added at a later
point.